reworked map pack widget mechanism
The map pack widget has been switched from my primitive custom data model (DirTreeModel) to an official file-system model (QFileSystemModel), because it is much more thoroughly written and it can be easily customized to display exactly what we want.
This should eliminate lags when you try to load a directory with huge number of maps.